Major Changes
Added -15V supply to OPA548, so we can get 0V. (Still needs to check on
bench if near 0V the OPA548 will consume any current from this rail);
Added a PDF with the schematic;
Added Arduino and Graphical 128x64 LCD
To be done:
Bead for +-15V analog power supply
Check if needs an op-amp after the DAC to drive the OPA548 (maybe we can
get a little trim?)
SDCard? Is this necessary? Are someone going to use this? (Maybe add and
doesn't build it?)
Output directly to Banana Plugs? Is this ok?