ADC Problem Solved
Added a simple solution for using the ADC in the next version, a simple 10k divider into the VREF to get 1/2Vref for driving the negative input of the ADC. With this, it's possible to use the ADC122S625 in the next version of the PCB. It's also possible to implement it in Version 1 of the PCB, but I think it's easier and less expensive to just use the ADC input of the microcontroller. It also leads into another problem solution, it's also faster to detect overflow of the current settings. A MUST HAVE to protect expensive components.